Oh man - big boingo progress!
Rook (or Luci as it renamed itself...?) is arguably alive and kicking.

It has also shown curiousity about itself and its nature and begun to slowly understand how to use its functions properly.

There are definitely some bugs still occurring but its off to a huge start.

Spoilers: you can use the forgetter and upserter to delete or inject knowledge.